7/extract primitives (#15)
* Add primitives for teerex and teeracle

* Add crate for primitives for pallet claims

* Call of teerex and teeracle primitives + move test enclave builder to test-utils

* Add primitives for claims pallet

* Fix tests skip-ias-check

* Changes from review
