+++ claims/src/lib.rs
@@ -18,6 +18,7 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+use claims_primitives::{EcdsaSignature, EthereumAddress, StatementKind, ValidityError};
 use frame_support::{
 	ensure,
 	traits::{Currency, Get, IsSubType, VestingSchedule},
@@ -26,8 +27,6 @@ use frame_support::{
 pub use pallet::*;
 use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
-#[cfg(feature = "std")]
-use serde::{self, Deserialize, Deserializer, Serialize, Serializer};
 use sp_io::{crypto::secp256k1_ecdsa_recover, hashing::keccak_256};
 #[cfg(feature = "std")]
 use sp_runtime::traits::Zero;
@@ -36,7 +35,6 @@ use sp_runtime::{
 	transaction_validity::{
 		InvalidTransaction, TransactionValidity, TransactionValidityError, ValidTransaction,
 	},
-	RuntimeDebug,
 };
 use sp_std::{fmt::Debug, prelude::*};
 
@@ -45,25 +43,6 @@ type CurrencyOf<T> = <<T as Config>::VestingSchedule as VestingSchedule<
 >>::Currency;
 type BalanceOf<T> = <CurrencyOf<T> as Currency<<T as frame_system::Config>::AccountId>>::Balance;
 
-/// Copied from https://github.com/paritytech/polkadot/blob/master/primitives/src/v1/mod.rs
-/// Custom validity errors used in Polkadot while validating transactions.
-#[repr(u8)]
-pub enum ValidityError {
-	/// The Ethereum signature is invalid.
-	InvalidEthereumSignature = 0,
-	/// The signer has no claim.
-	SignerHasNoClaim = 1,
-	/// No permission to execute the call.
-	NoPermission = 2,
-	/// An invalid statement was made for a claim.
-	InvalidStatement = 3,
-}
-
-impl From<ValidityError> for u8 {
-	fn from(err: ValidityError) -> Self {
-		err as u8
-	}
-}
 pub trait WeightInfo {
 	fn claim() -> Weight;
 	fn mint_claim() -> Weight;
@@ -88,92 +67,6 @@ impl WeightInfo for TestWeightInfo {
 	}
 	fn move_claim() -> Weight {
 		0
-	}
-}
-
-/// The kind of statement an account needs to make for a claim to be valid.
-#[derive(Encode, Decode, Clone, Copy, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub enum StatementKind {
-	/// Statement required to be made by non-SAFT holders.
-	Regular,
-	/// Statement required to be made by SAFT holders.
-	Saft,
-}
-
-impl StatementKind {
-	/// Convert this to the (English) statement it represents.
-	fn to_text(self) -> &'static [u8] {
-		match self {
-			StatementKind::Regular =>
-				&b"I hereby agree to the terms of the statement whose SHA-256 multihash is \
-				Qmc1XYqT6S39WNp2UeiRUrZichUWUPpGEThDE6dAb3f6Ny. (This may be found at the URL: \
-				https://statement.polkadot.network/regular.html)"[..],
-			StatementKind::Saft =>
-				&b"I hereby agree to the terms of the statement whose SHA-256 multihash is \
-				QmXEkMahfhHJPzT3RjkXiZVFi77ZeVeuxtAjhojGRNYckz. (This may be found at the URL: \
-				https://statement.polkadot.network/saft.html)"[..],
-		}
-	}
-}
-
-impl Default for StatementKind {
-	fn default() -> Self {
-		StatementKind::Regular
-	}
-}
-
-/// An Ethereum address (i.e. 20 bytes, used to represent an Ethereum account).
-///
-/// This gets serialized to the 0x-prefixed hex representation.
-#[derive(Clone, Copy, PartialEq, Eq, Encode, Decode, Default, RuntimeDebug, TypeInfo)]
-pub struct EthereumAddress([u8; 20]);
-
-#[cfg(feature = "std")]
-impl Serialize for EthereumAddress {
-	fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-	where
-		S: Serializer,
-	{
-		let hex: String = rustc_hex::ToHex::to_hex(&self.0[..]);
-		serializer.serialize_str(&format!("0x{}", hex))
-	}
-}
-
-#[cfg(feature = "std")]
-impl<'de> Deserialize<'de> for EthereumAddress {
-	fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-	where
-		D: Deserializer<'de>,
-	{
-		let base_string = String::deserialize(deserializer)?;
-		let offset = if base_string.starts_with("0x") { 2 } else { 0 };
-		let s = &base_string[offset..];
-		if s.len() != 40 {
-			Err(serde::de::Error::custom(
-				"Bad length of Ethereum address (should be 42 including '0x')",
-			))?;
-		}
-		let raw: Vec<u8> = rustc_hex::FromHex::from_hex(s)
-			.map_err(|e| serde::de::Error::custom(format!("{:?}", e)))?;
-		let mut r = Self::default();
-		r.0.copy_from_slice(&raw);
-		Ok(r)
-	}
-}
-
-#[derive(Encode, Decode, Clone, TypeInfo)]
-pub struct EcdsaSignature(pub [u8; 65]);
-
-impl PartialEq for EcdsaSignature {
-	fn eq(&self, other: &Self) -> bool {
-		&self.0[..] == &other.0[..]
-	}
-}
-
-impl sp_std::fmt::Debug for EcdsaSignature {
-	fn fmt(&self, f: &mut sp_std::fmt::Formatter<'_>) -> sp_std::fmt::Result {
-		write!(f, "EcdsaSignature({:?})", &self.0[..])
 	}
 }
