+++ primitives/teerex/Cargo.toml
@@ -1,0 +1,59 @@
+[package]
+name = "teerex-primitives"
+version = "0.1.0"
+authors = ["Integritee AG"]
+homepage = "https://integritee.network/"
+repository = "https://github.com/integritee-network/pallets/"
+license = "Apache-2.0"
+edition = "2021"
+resolver = "2"
+
+[dependencies]
+scale-info = { version = "1.0", default-features = false, features = ["derive"] }
+
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '2.3.1'
+
+[dependencies.ias-verify]
+default-features = false
+path = '../../teerex/ias-verify'
+
+[dependencies.frame-support]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+branch = "master"
+version = "4.0.0-dev"
+
+[dependencies.frame-system]
+default-features = false
+package = 'frame-system'
+git = "https://github.com/paritytech/substrate.git"
+branch = "master"
+version = "4.0.0-dev"
+
+[dependencies.sp-std]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+branch = "master"
+version = "4.0.0-dev"
+
+[dependencies.sp-core]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+branch = "master"
+version = "4.0.0-dev"
+
+[features]
+default = ['std']
+std = [
+    'codec/std',
+    'scale-info/std',
+    'frame-support/std',
+    'frame-system/std',
+    'ias-verify/std',
+    'sp-core/std',
+    'sp-std/std',
+]
